extract connection module
