+++ scylla/src/transport/cluster.rs
@@ -1,3 +1,4 @@
+use crate::connection::{Connection, PoolConfig, VerifiedKeyspaceName};
 /// Cluster manages up to date information and connections to database nodes
 use crate::frame::response::event::{Event, StatusChangeEvent};
 use crate::frame::value::ValueList;
@@ -5,8 +6,6 @@ use crate::prepared_statement::TokenCalculationError;
 use crate::sharding::Token;
 use crate::transport::host_filter::HostFilter;
 use crate::transport::{
-    connection::{Connection, VerifiedKeyspaceName},
-    connection_pool::PoolConfig,
     errors::QueryError,
     metadata::{Keyspace, Metadata, MetadataReader},
     node::Node,
