+++ scylla/src/transport/metadata.rs
@@ -1,8 +1,7 @@
+use crate::connection::{Connection, ConnectionConfig, NodeConnectionPool, PoolConfig, PoolSize};
 use crate::frame::response::event::Event;
 use crate::sharding::Token;
 use crate::statement::query::Query;
-use crate::transport::connection::{Connection, ConnectionConfig};
-use crate::transport::connection_pool::{NodeConnectionPool, PoolConfig, PoolSize};
 use crate::transport::errors::{DbError, QueryError};
 use crate::transport::host_filter::HostFilter;
 use crate::transport::node::resolve_contact_points;
