+++ scylla/src/transport/mod.rs
@@ -1,6 +1,4 @@
 pub(crate) mod cluster;
-pub(crate) mod connection;
-pub(crate) mod connection_pool;
 pub mod downgrading_consistency_retry_policy;
 pub mod execution_profile;
 pub mod host_filter;
