+++ scylla/src/transport/node.rs
@@ -2,11 +2,11 @@ use tokio::net::lookup_host;
 use tracing::warn;
 use uuid::Uuid;
 
+use crate::connection::Connection;
+use crate::connection::VerifiedKeyspaceName;
+use crate::connection::{NodeConnectionPool, PoolConfig};
 /// Node represents a cluster node along with it's data and connections
 use crate::sharding::{Sharder, Token};
-use crate::transport::connection::Connection;
-use crate::transport::connection::VerifiedKeyspaceName;
-use crate::transport::connection_pool::{NodeConnectionPool, PoolConfig};
 use crate::transport::errors::QueryError;
 
 use std::fmt::Display;
